requirements.txt
flask==1.1.1
requests==2.24.0
python-dotenv==0.14.0
Flask-Cors==3.0.9
gunicorn==20.0.4
